Add files via upload
Write a program that takes two numbers from the user and then calculates the sum of the product of the digits with the same positions.

Important points: 
The meaning of figures with the same position, for example, one, ten, etc. For example, in the numbers 32 and 14, the digits 2 and 4 have the same place.
The addition process continues until the last joint position. For example, the numbers 141 and 12 are calculated to the tenths.
Input 
The input consists of a line containing two numbers
�
mAnd
�
nThey are separated by a distance.
0
≤
�
,
�
≤
1
0
9
0≤m _n≤1 0 
9
 

Output 
The output consists of one line containing the sum.